APPSERV-59 initial update to MP Metrics 2.3

+++ appserver/payara-appserver-modules/microprofile/metrics/src/main/java/fish/payara/microprofile/metrics/impl/EWMA.java
@@ -159,6 +159,6 @@ public class EWMA {
      * @return the rate
      */
     public double getRate(TimeUnit rateUnit) {
-        return rate * (double) rateUnit.toNanos(1);
+        return rate * rateUnit.toNanos(1);
     }
 }
